# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión Académica",
        "description": "API REST para la gestión de cursos, estudiantes, docentes y materias",
        "contact": {
            "name": "Soporte API",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/docente-materia": {
            "get": {
                "tags": [
                    "Asignaciones Docente-Materia"
                ],
                "summary": "Obtener todas las asignaciones de docentes a materias",
                "operationId": "24fd09180e0bd106b96673860d1049cc",
                "responses": {
                    "200": {
                        "description": "Lista de asignaciones obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Asignaciones obtenidas correctamente"
                                        },
                                        "errors": {
                                            "type": "null"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "docentes_id_docentes": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "materias_id_materias": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "docente": {
                                                        "properties": {
                                                            "id_docente": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "nombre": {
                                                                "type": "string",
                                                                "example": "Juan"
                                                            },
                                                            "apellido": {
                                                                "type": "string",
                                                                "example": "Pérez"
                                                            },
                                                            "estado": {
                                                                "type": "string",
                                                                "example": "A"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "materia": {
                                                        "properties": {
                                                            "id_materia": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "nombre": {
                                                                "type": "string",
                                                                "example": "Matemáticas"
                                                            },
                                                            "estado": {
                                                                "type": "string",
                                                                "example": "A"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron asignaciones"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Asignaciones Docente-Materia"
                ],
                "summary": "Crear una nueva asignación de docente a materia",
                "operationId": "8d41ef54ca85f3552521f15eb04a025c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "docentes_id_docentes",
                                    "materias_id_materias"
                                ],
                                "properties": {
                                    "docentes_id_docentes": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "materias_id_materias": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Asignación creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Asignación creada correctamente"
                                        },
                                        "errors": {
                                            "type": "null"
                                        },
                                        "data": {
                                            "properties": {
                                                "docente_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "materia_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "409": {
                        "description": "El docente ya está asignado a esta materia"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}